+++ CHANGES.rst
@@ -11,6 +11,7 @@ in progress
   Other types will make the panel croak like ``InfluxDB Error: unsupported
   mean iterator type: *query.stringInterruptIterator`` or ``InfluxDB Error:
   not executed``.
+- Add TTS (The Things Stack) / TTN (The Things Network) decoder
 
 
 .. _kotori-0.27.0: